+++ source/geometry/box_geometry.h
@@ -6,11 +6,11 @@ class BoxGeometry : public Geometry {
 public:
     BoxGeometry(float3 half_extents);
 
-    std::optional<GeometryHit> raycast(const float3& origin, const float3& direction, float length) const override;
+    std::optional<GeometryHit> raycast(const float3& origin, const float3& direction, double length) const override;
 
     GeometrySample sample(const float2& random) const override;
 
-    float area() const override;
+    double area() const override;
 
 private:
     float3 m_half_extents;
