+++ source/geometry/sphere_geometry.h
@@ -4,14 +4,14 @@
 
 class SphereGeometry : public Geometry {
 public:
-    SphereGeometry(float radius);
+    SphereGeometry(double radius);
 
-    std::optional<GeometryHit> raycast(const float3& origin, const float3& direction, float length) const override;
+    std::optional<GeometryHit> raycast(const float3& origin, const float3& direction, double length) const override;
 
     GeometrySample sample(const float2& random) const override;
 
-    float area() const override;
+    double area() const override;
 
 private:
-    float m_radius;
+    double m_radius;
 };
